Add channel processing and transformation mixins with protocols
- Introduced `ChannelProcessingMixin` for audio signal processing methods including high-pass, low-pass filters, normalization, and more.
- Added `ChannelTransformMixin` for frequency transformation operations such as FFT, STFT, and Welch method.
- Created `protocols.py` to define common protocols for frame operations, including `BaseFrameProtocol`, `ProcessingFrameProtocol`, and `TransformFrameProtocol`.
- Updated `__init__.py` to include the new mixins for easier access.
