+++ wandas/frames/channel.py
@@ -1,30 +1,16 @@
 import logging
 from collections.abc import Iterator
 from pathlib import Path
-from typing import TYPE_CHECKING, Any, Callable, Optional, TypeVar, Union, cast
+from typing import Any, Callable, Optional, TypeVar, Union
 
 import dask
 import dask.array as da
 import matplotlib.pyplot as plt
 import numpy as np
 import soundfile as sf
+from dask.array.core import Array as DaArray
 from IPython.display import Audio, display
 from matplotlib.axes import Axes
-
-if TYPE_CHECKING:
-    from librosa._typing import (
-        _FloatLike_co,
-        _IntLike_co,
-        _PadModeSTFT,
-        _WindowSpec,
-    )
-
-    from .noct import NOctFrame
-    from .spectral import SpectralFrame
-    from .spectrogram import SpectrogramFrame
-
-
-from dask.array.core import Array as DaArray
 
 from wandas.utils.types import NDArrayReal
 
@@ -32,6 +18,7 @@ from ..core.base_frame import BaseFrame
 from ..core.metadata import ChannelMetadata
 from ..io.readers import get_file_reader
 from ..visualization.plotting import create_operation
+from .mixins import ChannelProcessingMixin, ChannelTransformMixin
 
 logger = logging.getLogger(__name__)
 
@@ -43,7 +30,9 @@ da_from_array = da.from_array  # type: ignore [unused-ignore]
 S = TypeVar("S", bound="BaseFrame[Any]")
 
 
-class ChannelFrame(BaseFrame[NDArrayReal]):
+class ChannelFrame(
+    BaseFrame[NDArrayReal], ChannelProcessingMixin, ChannelTransformMixin
+):
     """Channel-based data frame for handling audio signals and time series data.
 
     This frame represents channel-based data such as audio signals and time series data,
@@ -690,749 +679,6 @@ class ChannelFrame(BaseFrame[NDArrayReal]):
             data = data.squeeze(axis=1)
         sf.write(str(path), data, int(self.sampling_rate), format=format)
         logger.debug(f"Save complete: {path}")
-
-    def high_pass_filter(self, cutoff: float, order: int = 4) -> "ChannelFrame":
-        """Apply a high-pass filter to the signal.
-
-        Args:
-            cutoff: The cutoff frequency of the filter in Hz.
-            order: The order of the filter. Default is 4.
-
-        Returns:
-            A new ChannelFrame with the filtered signal.
-        """
-        logger.debug(
-            f"Setting up highpass filter: cutoff={cutoff}, order={order} (lazy)"
-        )
-        return self.apply_operation("highpass_filter", cutoff=cutoff, order=order)
-
-    def low_pass_filter(self, cutoff: float, order: int = 4) -> "ChannelFrame":
-        """Apply a low-pass filter to the signal.
-
-        Args:
-            cutoff: The cutoff frequency of the filter in Hz.
-            order: The order of the filter. Default is 4.
-
-        Returns:
-            A new ChannelFrame with the filtered signal.
-        """
-        logger.debug(
-            f"Setting up lowpass filter: cutoff={cutoff}, order={order} (lazy)"
-        )
-        return self.apply_operation("lowpass_filter", cutoff=cutoff, order=order)
-
-    def normalize(
-        self, target_level: float = -20, channel_wise: bool = True
-    ) -> "ChannelFrame":
-        """Normalize the signal level.
-
-        This method adjusts the signal amplitude to reach a target RMS level.
-
-        Args:
-            target_level: Target RMS level in dB. Default is -20.
-            channel_wise: If True, normalize each channel independently.
-                If False, apply the same scaling to all channels.
-
-        Returns:
-            A new ChannelFrame containing the normalized signal.
-        """
-        logger.debug(
-            f"Setting up normalize: target_level={target_level}, channel_wise={channel_wise} (lazy)"  # noqa: E501
-        )
-        return self.apply_operation(
-            "normalize", target_level=target_level, channel_wise=channel_wise
-        )
-
-    def a_weighting(self) -> "ChannelFrame":
-        """Apply A-weighting filter to the signal.
-
-        A-weighting adjusts the frequency response to approximate human hearing
-        perception, following IEC 61672-1:2013 standard.
-
-        Returns:
-            A new ChannelFrame with A-weighted signal.
-        """
-        return self.apply_operation("a_weighting")
-
-    def hpss_harmonic(
-        self,
-        kernel_size: Union[
-            "_IntLike_co", tuple["_IntLike_co", "_IntLike_co"], list["_IntLike_co"]
-        ] = 31,
-        power: float = 2,
-        margin: Union[
-            "_FloatLike_co",
-            tuple["_FloatLike_co", "_FloatLike_co"],
-            list["_FloatLike_co"],
-        ] = 1,
-        n_fft: int = 2048,
-        hop_length: Optional[int] = None,
-        win_length: Optional[int] = None,
-        window: "_WindowSpec" = "hann",
-        center: bool = True,
-        pad_mode: "_PadModeSTFT" = "constant",
-    ) -> "ChannelFrame":
-        """
-        Extract harmonic components using HPSS
-         (Harmonic-Percussive Source Separation).
-        """
-        return self.apply_operation(
-            "hpss_harmonic",
-            kernel_size=kernel_size,
-            power=power,
-            margin=margin,
-            n_fft=n_fft,
-            hop_length=hop_length,
-            win_length=win_length,
-            window=window,
-            center=center,
-            pad_mode=pad_mode,
-        )
-
-    def hpss_percussive(
-        self,
-        kernel_size: Union[
-            "_IntLike_co", tuple["_IntLike_co", "_IntLike_co"], list["_IntLike_co"]
-        ] = 31,
-        power: float = 2,
-        margin: Union[
-            "_FloatLike_co",
-            tuple["_FloatLike_co", "_FloatLike_co"],
-            list["_FloatLike_co"],
-        ] = 1,
-        n_fft: int = 2048,
-        hop_length: Optional[int] = None,
-        win_length: Optional[int] = None,
-        window: "_WindowSpec" = "hann",
-        center: bool = True,
-        pad_mode: "_PadModeSTFT" = "constant",
-    ) -> "ChannelFrame":
-        """
-        Extract percussive components using HPSS
-        (Harmonic-Percussive Source Separation).
-
-        This method separates the percussive (tonal) components from the signal.
-
-        Args:
-            kernel_size: Median filter size for HPSS.
-            power: Exponent for the Weiner filter used in HPSS.
-            margin: Margin size for the separation.
-
-        Returns:
-            A new ChannelFrame containing the harmonic components.
-        """
-        return self.apply_operation(
-            "hpss_percussive",
-            kernel_size=kernel_size,
-            power=power,
-            margin=margin,
-            n_fft=n_fft,
-            hop_length=hop_length,
-            win_length=win_length,
-            window=window,
-            center=center,
-            pad_mode=pad_mode,
-        )
-
-    def resampling(
-        self,
-        target_sr: float,
-        **kwargs: Any,
-    ) -> "ChannelFrame":
-        """
-        Resample audio data.
-
-        Parameters
-        ----------
-        target_sr : float
-            Target sampling rate (Hz)
-        resample_type : str, optional
-            Resampling method ('soxr_hq', 'linear', 'sinc', 'fft', etc.)
-        **kwargs : dict
-            Additional resampling parameters
-
-        Returns
-        -------
-        ChannelFrame
-            Resampled channel frame
-        """
-        return self.apply_operation(
-            "resampling",
-            target_sr=target_sr,
-            **kwargs,
-        )
-
-    def abs(self) -> "ChannelFrame":
-        """Calculate the absolute value of the signal.
-
-        Returns:
-            A new ChannelFrame containing the absolute values.
-        """
-        return self.apply_operation("abs")
-
-    def power(self, exponent: float = 2.0) -> "ChannelFrame":
-        """Calculate the power of the signal.
-
-        Args:
-            exponent: The exponent to raise the signal to. Default is 2.0.
-
-        Returns:
-            A new ChannelFrame containing the signal raised to the power.
-        """
-        return self.apply_operation("power", exponent=exponent)
-
-    def trim(
-        self,
-        start: float = 0,
-        end: Optional[float] = None,
-    ) -> "ChannelFrame":
-        """Trim the signal to specified time range.
-
-        Args:
-            start: Start time in seconds.
-            end: End time in seconds.
-
-        Returns:
-            A new ChannelFrame with trimmed signal.
-
-        Raises:
-            ValueError: If end time is before start time.
-        """
-        if end is None:
-            end = self.duration
-        if start > end:
-            raise ValueError("start must be less than end")
-        # Apply trim operation
-        return self.apply_operation("trim", start=start, end=end)
-
-    def rms_trend(
-        self,
-        frame_length: int = 2048,
-        hop_length: int = 512,
-        dB: bool = False,  # noqa: N803
-        Aw: bool = False,  # noqa: N803
-    ) -> "ChannelFrame":
-        """Calculate the RMS trend of the signal.
-
-        This method computes the root mean square value over sliding windows.
-
-        Args:
-            frame_length: The size of the sliding window in samples. Default is 2048.
-            hop_length: The hop length between windows in samples. Default is 512.
-            dB: Whether to return the RMS values in decibels. Default is False.
-            Aw: Whether to apply A-weighting. Default is False.
-
-        Returns:
-            A new ChannelFrame containing the RMS trend.
-        """
-        cf = self.apply_operation(
-            "rms_trend",
-            frame_length=frame_length,
-            hop_length=hop_length,
-            ref=[ch.ref for ch in self._channel_metadata],
-            dB=dB,
-            Aw=Aw,
-        )
-        cf.sampling_rate = self.sampling_rate / hop_length
-        return cf
-
-    def sum(self) -> "ChannelFrame":
-        """Sum all channels.
-
-        Returns:
-            A new ChannelFrame with summed signal.
-        """
-        return self.apply_operation("sum")
-
-    def mean(self) -> "ChannelFrame":
-        """Average all channels.
-
-        Returns:
-            A new ChannelFrame with averaged signal.
-        """
-        return self.apply_operation("mean")
-
-    def channel_difference(self, other_channel: Union[int, str] = 0) -> "ChannelFrame":
-        """Calculate channel differences relative to a reference channel.
-
-        Args:
-            other_channel: Reference channel index or label. Default is 0.
-
-        Returns:
-            A new ChannelFrame with channel differences.
-        """
-        if isinstance(other_channel, str):
-            return self.apply_operation(
-                "channel_difference", other_channel=self.label2index(other_channel)
-            )
-        return self.apply_operation("channel_difference", other_channel=other_channel)
-
-    def fft(self, n_fft: Optional[int] = None, window: str = "hann") -> "SpectralFrame":
-        """Compute Fast Fourier Transform.
-
-        Args:
-            n_fft: Number of FFT points. Default is next power of 2 of data length.
-            window: Window type. Default is "hann".
-
-        Returns:
-            A SpectralFrame containing the FFT results.
-        """
-        from ..processing import FFT
-        from .spectral import SpectralFrame
-
-        params = {"n_fft": n_fft, "window": window}
-        operation_name = "fft"
-        logger.debug(f"Applying operation={operation_name} with params={params} (lazy)")
-        from ..processing import create_operation
-
-        # Create operation instance
-        operation = create_operation(operation_name, self.sampling_rate, **params)
-        operation = cast("FFT", operation)
-        # Apply processing to data
-        spectrum_data = operation.process(self._data)
-
-        logger.debug(
-            f"Created new SpectralFrame with operation {operation_name} added to graph"
-        )
-
-        if n_fft is None:
-            is_even = spectrum_data.shape[-1] % 2 == 0
-            _n_fft = (
-                spectrum_data.shape[-1] * 2 - 2
-                if is_even
-                else spectrum_data.shape[-1] * 2 - 1
-            )
-        else:
-            _n_fft = n_fft
-
-        return SpectralFrame(
-            data=spectrum_data,
-            sampling_rate=self.sampling_rate,
-            n_fft=_n_fft,
-            window=operation.window,
-            label=f"Spectrum of {self.label}",
-            metadata={**self.metadata, "window": window, "n_fft": _n_fft},
-            operation_history=[
-                *self.operation_history,
-                {"operation": "fft", "params": {"n_fft": _n_fft, "window": window}},
-            ],
-            channel_metadata=self._channel_metadata,
-            previous=self,
-        )
-
-    def welch(
-        self,
-        n_fft: Optional[int] = None,
-        hop_length: Optional[int] = None,
-        win_length: int = 2048,
-        window: str = "hann",
-        average: str = "mean",
-    ) -> "SpectralFrame":
-        """Compute power spectral density using Welch's method.
-
-        Args:
-            n_fft: Number of FFT points. Default is 2048.
-            hop_length: Number of samples between successive frames.
-            Default is n_fft//4.
-            win_length: Length of window. Default is n_fft.
-            window: Window type. Default is "hann".
-            average: Method for averaging segments. Default is "mean".
-
-        Returns:
-            A SpectralFrame containing the power spectral density.
-        """
-        from ..processing import Welch
-        from .spectral import SpectralFrame
-
-        params = dict(
-            n_fft=n_fft or win_length,
-            hop_length=hop_length,
-            win_length=win_length,
-            window=window,
-            average=average,
-        )
-        operation_name = "welch"
-        logger.debug(f"Applying operation={operation_name} with params={params} (lazy)")
-        from ..processing import create_operation
-
-        # Create operation instance
-        operation = create_operation(operation_name, self.sampling_rate, **params)
-        operation = cast("Welch", operation)
-        # Apply processing to data
-        spectrum_data = operation.process(self._data)
-
-        logger.debug(
-            f"Created new SpectralFrame with operation {operation_name} added to graph"
-        )
-
-        return SpectralFrame(
-            data=spectrum_data,
-            sampling_rate=self.sampling_rate,
-            n_fft=operation.n_fft,
-            window=operation.window,
-            label=f"Spectrum of {self.label}",
-            metadata={**self.metadata, **params},
-            operation_history=[
-                *self.operation_history,
-                {"operation": "welch", "params": params},
-            ],
-            channel_metadata=self._channel_metadata,
-            previous=self,
-        )
-
-    def noct_spectrum(
-        self,
-        fmin: float,
-        fmax: float,
-        n: int = 3,
-        G: int = 10,  # noqa: N803
-        fr: int = 1000,
-    ) -> "NOctFrame":
-        """Compute N-octave band spectrum.
-
-        Args:
-            fmin: Minimum center frequency in Hz. Default is 20 Hz.
-            fmax: Maximum center frequency in Hz. Default is 20000 Hz.
-            n: Band division (1 for octave, 3 for 1/3 octave). Default is 3.
-            G: Reference gain in dB. Default is 10 dB.
-            fr: Reference frequency in Hz. Default is 1000 Hz.
-
-        Returns:
-            A NOctFrame containing the N-octave band spectrum.
-        """
-        from ..processing import NOctSpectrum
-        from .noct import NOctFrame
-
-        params = {"fmin": fmin, "fmax": fmax, "n": n, "G": G, "fr": fr}
-        operation_name = "noct_spectrum"
-        logger.debug(f"Applying operation={operation_name} with params={params} (lazy)")
-        from ..processing import create_operation
-
-        # Create operation instance
-        operation = create_operation(operation_name, self.sampling_rate, **params)
-        operation = cast("NOctSpectrum", operation)
-        # Apply processing to data
-        spectrum_data = operation.process(self._data)
-
-        logger.debug(
-            f"Created new SpectralFrame with operation {operation_name} added to graph"
-        )
-
-        return NOctFrame(
-            data=spectrum_data,
-            sampling_rate=self.sampling_rate,
-            fmin=fmin,
-            fmax=fmax,
-            n=n,
-            G=G,
-            fr=fr,
-            label=f"1/{n}Oct of {self.label}",
-            metadata={**self.metadata, **params},
-            operation_history=[
-                *self.operation_history,
-                {
-                    "operation": "noct_spectrum",
-                    "params": params,
-                },
-            ],
-            channel_metadata=self._channel_metadata,
-            previous=self,
-        )
-
-    def stft(
-        self,
-        n_fft: int = 2048,
-        hop_length: Optional[int] = None,
-        win_length: Optional[int] = None,
-        window: str = "hann",
-    ) -> "SpectrogramFrame":
-        """Compute Short-Time Fourier Transform.
-
-        Args:
-            n_fft: Number of FFT points. Default is 2048.
-            hop_length: Number of samples between successive frames.
-            Default is n_fft//4.
-            win_length: Length of window. Default is n_fft.
-            window: Window type. Default is "hann".
-
-        Returns:
-            A SpectrogramFrame containing the STFT results.
-        """
-        from ..processing import STFT, create_operation
-        from .spectrogram import SpectrogramFrame
-
-        # Set hop length and window length
-        _hop_length = hop_length if hop_length is not None else n_fft // 4
-        _win_length = win_length if win_length is not None else n_fft
-
-        params = {
-            "n_fft": n_fft,
-            "hop_length": _hop_length,
-            "win_length": _win_length,
-            "window": window,
-        }
-        operation_name = "stft"
-        logger.debug(f"Applying operation={operation_name} with params={params} (lazy)")
-
-        # Create operation instance
-        operation = create_operation(operation_name, self.sampling_rate, **params)
-        operation = cast("STFT", operation)
-
-        # Apply processing to data
-        spectrogram_data = operation.process(self._data)
-
-        logger.debug(
-            f"Created new SpectrogramFrame with operation {operation_name} added to graph"  # noqa: E501
-        )
-
-        # Create new instance
-        return SpectrogramFrame(
-            data=spectrogram_data,
-            sampling_rate=self.sampling_rate,
-            n_fft=n_fft,
-            hop_length=_hop_length,
-            win_length=_win_length,
-            window=window,
-            label=f"stft({self.label})",
-            metadata=self.metadata,
-            operation_history=self.operation_history,
-            channel_metadata=self._channel_metadata,
-        )
-
-    def coherence(
-        self,
-        n_fft: int = 2048,
-        hop_length: Optional[int] = None,
-        win_length: Optional[int] = None,
-        window: str = "hann",
-        detrend: str = "constant",
-    ) -> "SpectralFrame":
-        """Compute magnitude squared coherence.
-
-        Args:
-            n_fft: Number of FFT points. Default is 2048.
-            hop_length: Number of samples between successive frames.
-            Default is n_fft//4.
-            win_length: Length of window. Default is n_fft.
-            window: Window type. Default is "hann".
-            detrend: Detrending method. Options: "constant", "linear", None.
-
-        Returns:
-            A SpectralFrame containing the magnitude squared coherence.
-        """
-        from ..processing import Coherence, create_operation
-        from .spectral import SpectralFrame
-
-        params = {
-            "n_fft": n_fft,
-            "hop_length": hop_length,
-            "win_length": win_length,
-            "window": window,
-            "detrend": detrend,
-        }
-        operation_name = "coherence"
-        logger.debug(f"Applying operation={operation_name} with params={params} (lazy)")
-
-        # Create operation instance
-        operation = create_operation(operation_name, self.sampling_rate, **params)
-        operation = cast("Coherence", operation)
-
-        # Apply processing to data
-        coherence_data = operation.process(self._data)
-
-        logger.debug(
-            f"Created new SpectralFrame with operation {operation_name} added to graph"  # noqa: E501
-        )
-        # Create new channel metadata
-        channel_metadata = []
-        for in_ch in self._channel_metadata:
-            for out_ch in self._channel_metadata:
-                meta = ChannelMetadata()
-                meta.label = f"$\\gamma_{{{in_ch.label}, {out_ch.label}}}$"
-                meta.unit = ""
-                meta.ref = 1
-                meta["metadata"] = dict(
-                    in_ch=in_ch["metadata"], out_ch=out_ch["metadata"]
-                )
-                channel_metadata.append(meta)
-
-        # Create new instance
-        return SpectralFrame(
-            data=coherence_data,
-            sampling_rate=self.sampling_rate,
-            n_fft=operation.n_fft,
-            window=operation.window,
-            label=f"Coherence of {self.label}",
-            metadata={**self.metadata, **params},
-            operation_history=[
-                *self.operation_history,
-                {"operation": operation_name, "params": params},
-            ],
-            channel_metadata=channel_metadata,
-            previous=self,
-        )
-
-    def csd(
-        self,
-        n_fft: int = 2048,
-        hop_length: Optional[int] = None,
-        win_length: Optional[int] = None,
-        window: str = "hann",
-        detrend: str = "constant",
-        scaling: str = "spectrum",
-        average: str = "mean",
-    ) -> "SpectralFrame":
-        """Compute cross-spectral density matrix.
-
-        Args:
-            n_fft: Number of FFT points. Default is 2048.
-            hop_length: Number of samples between successive frames.
-            Default is n_fft//4.
-            win_length: Length of window. Default is n_fft.
-            window: Window type. Default is "hann".
-            detrend: Detrending method. Options: "constant", "linear", None.
-            scaling: Scaling method. Options: "spectrum", "density".
-            average: Method for averaging segments. Default is "mean".
-
-        Returns:
-            A SpectralFrame containing the cross-spectral density matrix.
-        """
-        from ..processing import CSD, create_operation
-        from .spectral import SpectralFrame
-
-        params = {
-            "n_fft": n_fft,
-            "hop_length": hop_length,
-            "win_length": win_length,
-            "window": window,
-            "detrend": detrend,
-            "scaling": scaling,
-            "average": average,
-        }
-        operation_name = "csd"
-        logger.debug(f"Applying operation={operation_name} with params={params} (lazy)")
-
-        # Create operation instance
-        operation = create_operation(operation_name, self.sampling_rate, **params)
-        operation = cast("CSD", operation)
-
-        # Apply processing to data
-        csd_data = operation.process(self._data)
-
-        logger.debug(
-            f"Created new SpectralFrame with operation {operation_name} added to graph"  # noqa: E501
-        )
-        # Create new channel metadata
-        channel_metadata = []
-        for in_ch in self._channel_metadata:
-            for out_ch in self._channel_metadata:
-                meta = ChannelMetadata()
-                meta.label = f"{operation_name}({in_ch.label}, {out_ch.label})"
-                meta.unit = ""
-                meta.ref = 1
-                meta["metadata"] = dict(
-                    in_ch=in_ch["metadata"], out_ch=out_ch["metadata"]
-                )
-                channel_metadata.append(meta)
-
-        # Create new instance
-        return SpectralFrame(
-            data=csd_data,
-            sampling_rate=self.sampling_rate,
-            n_fft=operation.n_fft,
-            window=operation.window,
-            label=f"$C_{{{in_ch.label}, {out_ch.label}}}$",
-            metadata={**self.metadata, **params},
-            operation_history=[
-                *self.operation_history,
-                {"operation": operation_name, "params": params},
-            ],
-            channel_metadata=channel_metadata,
-            previous=self,
-        )
-
-    def transfer_function(
-        self,
-        n_fft: int = 2048,
-        hop_length: Optional[int] = None,
-        win_length: Optional[int] = None,
-        window: str = "hann",
-        detrend: str = "constant",
-        scaling: str = "spectrum",
-        average: str = "mean",
-    ) -> "SpectralFrame":
-        """Compute transfer function matrix.
-
-        The transfer function characterizes the signal transmission properties
-        between channels in the frequency domain, representing the input-output
-        relationship of a system.
-
-        Args:
-            n_fft: Number of FFT points. Default is 2048.
-            hop_length: Number of samples between successive frames.
-            Default is n_fft//4.
-            win_length: Length of window. Default is n_fft.
-            window: Window type. Default is "hann".
-            detrend: Detrending method. Options: "constant", "linear", None.
-            scaling: Scaling method. Options: "spectrum", "density".
-            average: Method for averaging segments. Default is "mean".
-
-        Returns:
-            A SpectralFrame containing the transfer function matrix.
-        """
-        from ..processing import TransferFunction, create_operation
-        from .spectral import SpectralFrame
-
-        params = {
-            "n_fft": n_fft,
-            "hop_length": hop_length,
-            "win_length": win_length,
-            "window": window,
-            "detrend": detrend,
-            "scaling": scaling,
-            "average": average,
-        }
-        operation_name = "transfer_function"
-        logger.debug(f"Applying operation={operation_name} with params={params} (lazy)")
-
-        # Create operation instance
-        operation = create_operation(operation_name, self.sampling_rate, **params)
-        operation = cast("TransferFunction", operation)
-
-        # Apply processing to data
-        tf_data = operation.process(self._data)
-
-        logger.debug(
-            f"Created new SpectralFrame with operation {operation_name} added to graph"  # noqa: E501
-        )
-        # Create new channel metadata
-        channel_metadata = []
-        for in_ch in self._channel_metadata:
-            for out_ch in self._channel_metadata:
-                meta = ChannelMetadata()
-                meta.label = f"$H_{{{in_ch.label}, {out_ch.label}}}$"
-                meta.unit = ""
-                meta.ref = 1
-                meta["metadata"] = dict(
-                    in_ch=in_ch["metadata"], out_ch=out_ch["metadata"]
-                )
-                channel_metadata.append(meta)
-
-        # Create new instance
-        return SpectralFrame(
-            data=tf_data,
-            sampling_rate=self.sampling_rate,
-            n_fft=operation.n_fft,
-            window=operation.window,
-            label=f"Transfer function of {self.label}",
-            metadata={**self.metadata, **params},
-            operation_history=[
-                *self.operation_history,
-                {"operation": operation_name, "params": params},
-            ],
-            channel_metadata=channel_metadata,
-            previous=self,
-        )
 
     def _get_additional_init_kwargs(self) -> dict[str, Any]:
         """Provide additional initialization arguments required for ChannelFrame."""
