+++ wandas/frames/mixins/__init__.py
@@ -1,0 +1,6 @@
+"""Channel frame mixins module."""
+
+from .channel_processing_mixin import ChannelProcessingMixin
+from .channel_transform_mixin import ChannelTransformMixin
+
+__all__ = ["ChannelProcessingMixin", "ChannelTransformMixin"]
